Move Hexify type into util module
Move the Hexify helper type into the util module -- as less opinionated
location for this reusable functionality.
